Proyecto -As You Wish (All Version)/ProyectosUNAB (Final Version)/AsYouWish_Serv/requirements.txt
django
djangorestframework
gunicorn
coverage
django-cors-headers
django-filter
mysql-connector-python